[proj-104] feat: real estate에 저장되는 deposit, price의 타입을 Long으로 변경

+++ src/main/java/com/omnm/hanasset/loan/service/LoanService.java
@@ -64,8 +64,8 @@ public class LoanService {
                     property.getIsHousingFraudVictim(),
                     housingType.getExclusiveAreaSize().intValue(),
                     realEstate.getType(),
-                    realEstate.getDeposit() / 10000,
-                    realEstate.getPrice() / 10000);
+                    (int) (realEstate.getDeposit() / 10000),
+                    (int) (realEstate.getPrice() / 10000));
 
             for (Loan loan : availableLoans) {
                 LoanInfoDTO loanInfoDTO = loanMapper.loanToInfoDTO(loan);
